Rewrite to hide the body from SPARK.
Implementing some features, such as Reference and Constant_Reference, purely
in SPARK does not seem feasible due to some accessibility VCs when built on top
of containers which take an aliased container parameters. Instead, let's switch
approach to the SPARK library which defines the interface and a formal model in
SPARK, but with the implementation in Ada.