Refactor model of Has_Element and fix postcondition of Insert_Child
Previously, the formal model depended on the declaration of Has_Element on type Tree,
whereas all the dependency is the other way around for all other Tree operations
(i.e. the Tree operation depends on the formal model in their contracts).

Additionally, the postcondition of Insert_Child was not correctly stating that a new
child element had been inserted.

The combination of both of these were leading to some unproved checks, which are now
fixed.