Rework postcondition of Insert_Parent and add output parameter for the new node
The previous postcondition was not correct and caused issues with proof.
It's also very useful in many cases to have Insert_Parent output a cursor
to the new node, as quite often the new node is used in some way. It also
helps simplify the postcondition by being able to reference the new node.