Rework Insert_Parent spec.
This removes the New_Node output parameter as it turns out that it's not
always needed by the caller, and if it is then it's easy enough to get
it by calling the Parent function afterwards.

The postcondition of Insert_Parent is also reworked to simplify the properties
it proves and fix some correctness issues.