Fix inconsistencies in postcondition of Insert_Parent
There was an inconsistency that lead to a contradiction in the part of the
postcondition that asserts that existing children of the target node are
unchanged. This lead to the provers being able to erroneously prove anything
at all (even Assert(False)). This is now fixed.

There was also a missing condition that asserts that the new node is a leaf
node, causing the proof tests to (correctly) fail when the above contradiction
was removed.

It also turns out that Inline_For_Proof was a bit detrimental, so it is also
removed.