WIP: Rework functional trees and update contracts of Unbounded_Multiway_Trees
This rework allows the functional tree to be used as the formal model for the
unbounded tree. The main part of the unbounded tree's contracts is the mapping
of cursors to nodes in the formal model, which are represented the path to the
node from the root of the tree.

This commit is the first working version that passes the proof tests, but some
subprograms are just stubs for now and need to be properly implemented.