Implement functional trees for use as the formal model (#3)
* Improve postcondition of Insert_Child and Model

* Add In_Subtree and In_Branch functions

* Add formal model function for Root

* Implement new package Stree.Functional_Multiway_Trees

This implements a functional version of the tree container, following
the same model as the SPARK functional containers library.

* Add unit tests for functional Add and Add_Parent

* proof_tests: enable --proof-warnings

* Add proof tests for functional multiway tree

* Implement Length and reformat some contracts

* WIP: Rework functional trees and update contracts of Unbounded_Multiway_Trees

This rework allows the functional tree to be used as the formal model for the
unbounded tree. The main part of the unbounded tree's contracts is the mapping
of cursors to nodes in the formal model, which are represented the path to the
node from the root of the tree.

This commit is the first working version that passes the proof tests, but some
subprograms are just stubs for now and need to be properly implemented.

* Update proof tests to reflect new contracts

* Update unit tests to reflect changes to functional trees

* Add missing comments for several property functions

* Add a comment to explain the mapping between cursors and paths

* Replace construction functions with property functions in contracts

This is more efficient in proof.

* Implement Paths function

* Simplify implementation of Insert

* Implement M_Path

* Implement M_Cursor

* Remove M_Cursor

It turns out that the reverse mapping of paths to cursors is not needed,
and its removal makes the proofs more efficient.

* Implement Mapping_Preserved

* Implement Mapping_Preserved_Except_Subtree

* Implement Same_Mapping_Except

* Implement Same_Mapping_Except_Subtree

* Relax preconditions of functions Parent and Child

* Minor corrections to comments

* Fix postcondition of Insert

* Implement iteration on functional trees (Iterable_Tree)

* Add comments to several subprograms' contracts

* Add copyright header

* Use Copy_Element correctly in contracts

* Add missing comments for iteration-related private subprograms

* Add comment explaining why Choose always returns a leaf