Avoid accidental recursion in contracts
The implementation of certain subprograms made use of other functions
which called the first subprogram in its contracts, leading to accidental
recursion when contracts are enabled. For example, the implementation of
Model called Next, but next called back to Model in its postcondition
(indirectly via function Positions). This resulted in unbounded recursion
when postcondition assertions were enabled.

This commit adds variants of certain subprograms which don't have contracts,
so they can be safely called in the implementation of other subprograms.
For example, Next_Impl implements the logic of Next, but has no contracts
so it can be safely called by Model.