Corrections to handling Element_Type in contracts
In the functional trees, Element_Logic_Equal is used to compare elements
and Copy_Element must be used when comparing an element parameter to ensure
correct rejection of access types.

In Unbounded_Multiway_Trees, Equivalent_Elements is removed and replaced with
operator "=" to ensure equality is always checked correctly.

The postcondition of Constant_Reference is also fixed to use Element_Logic_Equal
appropriately.