Strengthen contracts for Insert_Child, Insert_Parent, and Delete
The contracts for Insert_Parent stated how the mapping was preserved
from the old and new containers, but said nothing in the other direction.
This is now corrected and a new property function Subtree_Mapping_Shifted_Up
is added.

Also, the provers found it difficult to prove that the element associated with
each cursor is preserved for Insert_Parent, so a new property function
Elements_Preserved is added to prove this property.

The precondition for Direction is also corrected to require that the target
node is not the root node, since otherwise it violates the precondition of
M.Way_From_Parent in the call from Direction's postcondition.